backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.32.1
mysql-connector-python==9.1.0
pydantic==2.10.4
pydantic[email]==2.10.4
python-multipart==0.0.20 